결과 데이터 처리 방법 변경 컨트롤러에 적용

+++ src/main/java/com/example/demo/service/ArticleService.java
@@ -33,11 +33,17 @@ public class ArticleService {
 		return this.articleDao.getArticleById(id);
 	}
 
-	public void modifyArticle(Article foundArticle, String title, String body) {
+	public void modifyArticle(String writer, Article foundArticle, String title, String body) {
+		if (!writer.equals(foundArticle.getWriter())) {
+			throw new IllegalStateException("작성자와 로그인 사용자가 일치하지 않습니다.");
+		}
 		this.articleDao.updateArticle(foundArticle, title, body);
 	}
 
-	public void deleteArticle(Article foundArticle) {
+	public void deleteArticle(String writer, Article foundArticle) {
+		if (!writer.equals(foundArticle.getWriter())) {
+			throw new IllegalStateException("작성자와 로그인 사용자가 일치하지 않습니다.");
+		}
 		this.articleDao.deleteArticle(foundArticle);
 	}
